x/cainiao/types/message_add_order_test.go
package types

import (
	"testing"

	"github.com/cosmonaut/cainiao/testutil/sample"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgAddOrder_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgAddOrder
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgAddOrder{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgAddOrder{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
